optimization_engine/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
ortools==9.10.4067
supabase==2.7.4
httpx==0.27.2
apscheduler==3.10.4
python-dotenv==1.0.1
pydantic==2.8.2
